requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
dblpy==0.3.4
discord.py==1.3.3
idna==2.9
multidict==4.7.6
numpy==1.19.0
pandas==1.0.5
psutil==5.7.0
python-dateutil==2.8.1
pytz==2020.1
ratelimiter==1.2.0.post0
requests==2.24.0
six==1.15.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
